change types

+++ C++/Navio/Common/MPU9250.h
@@ -15,12 +15,12 @@ public:
   void update() override;
 
 private:
-  uint32_t WriteReg(uint8_t WriteAddr, uint8_t WriteData);
-  uint32_t ReadReg(uint8_t ReadAddr);
-  void ReadRegs(uint8_t ReadAddr, uint8_t* ReadBuf, uint32_t Bytes);
+  uint8_t WriteReg(uint8_t WriteAddr, uint8_t WriteData);
+  uint8_t ReadReg(uint8_t ReadAddr);
+  void ReadRegs(uint8_t ReadAddr, uint8_t* ReadBuf, uint8_t Bytes);
 
-  uint32_t set_gyro_scale(int scale);
-  uint32_t set_acc_scale(int scale);
+  uint8_t set_gyro_scale(uint8_t scale);
+  uint8_t set_acc_scale(uint8_t scale);
 
   void calib_acc();
   void calib_mag();
