+++ C++/Navio/Navio2/RCOutput_Navio2.h
@@ -8,10 +8,10 @@ class RCOutput_Navio2 : public RCOutput
 public:
   inline explicit RCOutput_Navio2();
 
-  inline bool initialize(const uint32_t& channel) override;
-  inline bool enable(const uint32_t& channel) override;
-  inline bool setFrequency(const uint32_t& channel, const uint32_t& frequency) override;
-  inline bool setDutyCycle(const uint32_t& channel, const double& period_us) override;
+  inline bool initialize(const size_t& channel) override;
+  inline bool enable(const size_t& channel) override;
+  inline bool setFrequency(const size_t& channel, const size_t& frequency) override;
+  inline bool setDutyCycle(const size_t& channel, const double& period_us) override;
 
 private:
   PWM pwm_;
@@ -21,22 +21,22 @@ inline RCOutput_Navio2::RCOutput_Navio2()
 {
 }
 
-inline bool RCOutput_Navio2::initialize(const uint32_t& channel)
+inline bool RCOutput_Navio2::initialize(const size_t& channel)
 {
   return pwm_.init(channel);
 }
 
-inline bool RCOutput_Navio2::enable(const uint32_t& channel)
+inline bool RCOutput_Navio2::enable(const size_t& channel)
 {
   return pwm_.enable(channel);
 }
 
-inline bool RCOutput_Navio2::setFrequency(const uint32_t& channel, const uint32_t& frequency)
+inline bool RCOutput_Navio2::setFrequency(const size_t& channel, const size_t& frequency)
 {
   return pwm_.setPeriod(channel, frequency);
 }
 
-inline bool RCOutput_Navio2::setDutyCycle(const uint32_t& channel, const double& period_us)
+inline bool RCOutput_Navio2::setDutyCycle(const size_t& channel, const double& period_us)
 {
   return pwm_.setDutyCycle(channel, period_us / 1000);
 }
